+++ src/neo-express/Commands/ContractCommand.Get.cs
@@ -1,10 +1,7 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
-using NeoExpress.Models;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using System;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace NeoExpress.Commands
@@ -24,31 +21,31 @@ namespace NeoExpress.Commands
             [Option]
             private bool Overwrite { get; }
 
-            async Task<int> OnExecuteAsync(CommandLineApplication app, IConsole console)
-            {
-                try
-                {
-                    var (chain, filename) = Program.LoadExpressChain(Input);
-                    var contract = chain.GetContract(Contract);
-                    if (contract == null)
-                    {
-                        throw new Exception($"Contract {Contract} not found.");
-                    }
+            // async Task<int> OnExecuteAsync(CommandLineApplication app, IConsole console)
+            // {
+            //     try
+            //     {
+            //         var (chain, filename) = Program.LoadExpressChain(Input);
+            //         var contract = chain.GetContract(Contract);
+            //         if (contract == null)
+            //         {
+            //             throw new Exception($"Contract {Contract} not found.");
+            //         }
 
-                    var uri = chain.GetUri();
-                    var result = await NeoRpcClient.GetContractState(uri, contract.Hash).ConfigureAwait(false);
-                    console.WriteResult(result);
+            //         var uri = chain.GetUri();
+            //         var result = await NeoRpcClient.GetContractState(uri, contract.Hash).ConfigureAwait(false);
+            //         console.WriteResult(result);
 
-                    chain.SaveContract(contract, filename, console, Overwrite);
-                    return 0;
-                }
-                catch (Exception ex)
-                {
-                    console.WriteError(ex.Message);
-                    app.ShowHelp();
-                    return 1;
-                }
-            }
+            //         chain.SaveContract(contract, filename, console, Overwrite);
+            //         return 0;
+            //     }
+            //     catch (Exception ex)
+            //     {
+            //         console.WriteError(ex.Message);
+            //         app.ShowHelp();
+            //         return 1;
+            //     }
+            // }
         }
     }
 }
